remove sha
