+++ docs/reference/experiment-config.txt
@@ -676,162 +676,6 @@ large-scale experiments with hundreds or thousands of trials.
    which to initialize weights. At most one of ``source_trial_id`` or
    ``source_checkpoint_uuid`` should be set.
 
-Adaptive (Simple)
-=================
-
-.. warning::
-
-   Adaptive (Simple) is deprecated and will be removed in a future
-   release. We recommend using the state-of-the-art :ref:`Adaptive
-   (ASHA) <experiment-configuration-searcher-adaptive>` searcher.
-
-The ``adaptive_simple`` search method is a simpler interface to the
-:ref:`adaptive <experiment-configuration-searcher-adaptive>` search
-method described above. ``adaptive_simple`` is designed to be simpler to
-configure for most applications of hyperparameter search.
-
-**Required Fields**
-
-``metric``
-   The name of the validation metric used to evaluate the performance of
-   a hyperparameter configuration.
-
-``max_length``:
-   The maximum training length of any one trial, in terms of records,
-   batches, or epochs (see :ref:`Training Units
-   <experiment-configuration_training_units>`). The vast majority of
-   trials will be stopped early, and thus only a small fraction of
-   trials will actually be trained for this long. This quantity is
-   domain-specific and should roughly reflect the length of training
-   needed for the model to converge on the data set.
-
-``max_trials``
-   The number of trials, i.e., hyperparameter configurations, to
-   evaluate.
-
-**Optional Fields**
-
-``smaller_is_better``
-   Whether to minimize or maximize the metric defined above. The default
-   value is ``true`` (minimize).
-
-``mode``
-   How aggressively to perform early stopping. There are three modes:
-   ``aggressive``, ``standard``, and ``conservative``; the default is
-   ``standard``.
-
-   These modes differ in the degree to which early-stopping is used. In
-   ``aggressive`` mode, the searcher quickly stops underperforming
-   trials, which enables the searcher to explore more hyperparameter
-   configurations, but at the risk of discarding a configuration too
-   soon. On the other end of the spectrum, ``conservative`` mode
-   performs significantly less downsampling, but as a consequence does
-   not explore as many configurations given the same budget. We
-   recommend using either ``aggressive`` or ``standard`` mode.
-
-``divisor``
-   The fraction of trials to keep at each rung, and also determines the
-   training length for each rung. The default setting is ``4``; only
-   advanced users should consider changing this value.
-
-``max_rungs``
-   The maximum number of times we evaluate intermediate results for a
-   trial and terminate poorly performing trials. The default value is
-   ``5``; only advanced users should consider changing this value.
-
-``source_trial_id``
-   If specified, the weights of *every* trial in the search will be
-   initialized to the most recent checkpoint of the given trial ID. This
-   will fail if the source trial's model architecture is inconsistent
-   with the model architecture of any of the trials in this experiment.
-
-``source_checkpoint_uuid``
-   Like ``source_trial_id``, but specifies an arbitrary checkpoint from
-   which to initialize weights. At most one of ``source_trial_id`` or
-   ``source_checkpoint_uuid`` should be set.
-
-Adaptive (Advanced)
-===================
-
-.. warning::
-
-   Adaptive (Advanced) is deprecated and will be removed in a future
-   release. We recommend using the state-of-the-art :ref:`Adaptive
-   (ASHA) <experiment-configuration-searcher-adaptive>` searcher.
-
-The ``adaptive`` search method is a theoretically principled and
-empirically state-of-the-art method that adaptively allocates resources
-to promising hyperparameter configurations while quickly eliminating
-poor ones.
-
-**Required Fields**
-
-``metric``
-   The name of the validation metric used to evaluate the performance of
-   a hyperparameter configuration.
-
-``max_length``
-   The maximum training length of any one trial, in terms of records,
-   batches or epochs (see :ref:`Training Units
-   <experiment-configuration_training_units>`). The vast majority of
-   trials will be stopped early, and thus only a small fraction of
-   trials will actually be trained for this long. We suggest setting
-   this to a multiple of ``divisor^(max_rungs-1)``, which is 4^(5-1) =
-   256 with the default values.
-
-.. _experiment-configuration_adaptive-advanced-budget:
-
-``budget``
-   The total training length across all trials, in terms of the same
-   unit (see :ref:`Training Units
-   <experiment-configuration_training_units>`) as ``max_length``. We
-   suggest setting this to be a multiple of ``max_length``, which
-   implies interpreting this subfield as the effective number of
-   complete trials to evaluate. Note that some trials might be
-   in-progress when this budget is exhausted; adaptive search will allow
-   these to complete.
-
-**Optional Fields**
-
-``smaller_is_better``
-   Whether to minimize or maximize the metric defined above. The default
-   value is ``true`` (minimize).
-
-``mode``
-   How aggressively to perform early stopping. There are three modes:
-   ``aggressive``, ``standard``, and ``conservative``; the default is
-   ``standard``.
-
-   These modes differ in the degree to which early-stopping is used. In
-   ``aggressive`` mode, the searcher quickly stops underperforming
-   trials, which enables the searcher to explore more hyperparameter
-   configurations, but at the risk of discarding a configuration too
-   soon. On the other end of the spectrum, ``conservative`` mode
-   performs significantly less downsampling, but as a consequence does
-   not explore as many configurations given the same budget. We
-   recommend using either ``aggressive`` or ``standard`` mode.
-
-``divisor``
-   The fraction of trials to keep at each rung, and also determines the
-   training length for each rung. The default setting is ``4``; only
-   advanced users should consider changing this value.
-
-``max_rungs``
-   The maximum number of times we evaluate intermediate results for a
-   trial and terminate poorly performing trials. The default value is
-   ``5``; only advanced users should consider changing this value.
-
-``source_trial_id``
-   If specified, the weights of *every* trial in the search will be
-   initialized to the most recent checkpoint of the given trial ID. This
-   will fail if the source trial's model architecture is inconsistent
-   with the model architecture of any of the trials in this experiment.
-
-``source_checkpoint_uuid``
-   Like ``source_trial_id``, but specifies an arbitrary checkpoint from
-   which to initialize weights. At most one of ``source_trial_id`` or
-   ``source_checkpoint_uuid`` should be set.
-
 PBT
 ===
