+++ e2e_tests/tests/fixtures/no_op/adaptive.yaml
@@ -12,15 +12,11 @@ hyperparameters:
     maxval: 0.9
   metrics_sigma: 0
 searcher:
-  name: adaptive
+  name: adaptive_asha
   metric: validation_error
+  max_trials: 30
   max_length:
     batches: 640
-  budget:
-    batches: 6000
-  mode: standard
-  max_rungs: 3
-  divisor: 4
 reproducibility:
   experiment_seed: 999
 max_restarts: 0
